move routing data to its own directory
